Fixed an error in the communication with the broker. Closes #8.
The `cb_brokern_on_out` used to read and parse the data from the broker as soon as it got it, without making sure it got an entire message. This is now fixed by treating the broker output as a continuous stream and only taking data from it when we encounter a newline character.