Lots of improvements: - new gdb commands such as rln, bbt, bx, cc and patch - better synchronisation - support external pid, mapping These improvements are detailed below.
The new gdb commands are:
- rln: ask IDA Pro to convert an address into a symbol
- bbt: beautiful backtrace by executing "bt" and retrieving
       symbols from IDA Pro for each element of the bt
- bx: similar to "x" but supports a symbol resolved by IDA
- cc: continue to current cursor in IDA Pro (set a breakpoint,
      continue and remove the bp)
- patch: patch memory in IDA based on live information in gdb

The communication with the broker as been improved.
Sometimes the whole JSON is not received at once.
We cache it until we receive the ending "}".

Added support for importing a context to the gdb plugin.
In some scenario such as debugging embedded devices over
serial, gdb is not aware of the PID and cannot access
/proc/<pid>/maps. In this case, we import a different file
than sync.py in gdb which is quite simple and import sync.py.
The only difference is we create a context that is passed
to Sync() when instanciating the object. It allows overriding
some fields such as the pid or mappings but other could be
added if required.

```
from sync import *

if __name__ == "__main__":

    ctx = { "pid": 200,
            "mappings": [0x08048000, 0x0B74ae20, 0x03702e20, "asa924-k8.bin|lina"]
    }
    //...
    sync = Sync(HOST, ctx=ctx)
    //...
```

A few other small fixes such as:
- better parsing of config file at several places
- more logging in IDA output window
- fixed some polling error that appeared in gdb when syncing failed