Add new configuration option use_tmp_logging_file to gdb extension.
The reason this was added was that the use of a temporary file to
capture the gdb output does not work in my case - namely when using
the Qt Creator debugging frontend. The output was messed up and
the proc map could not be parsed correctly.

With the new option the output to temporary files can be deactivated
which worked in my case.