[ext_gdb] use a different input for deadlock prevention instead of new_thread event
As reported in https://github.com/bootleg/ret-sync/issues/66 , new_thread event may not be available for some packages of GDB. Remove dependency on it. Use a check on threading.main_thread().is_alive() instead to detect the termination of the interpreter.

Thanks @sn0ot and @agantet for reporting the issue.