fixed a bug which causes dispatcher crash
with Python 3.10 and trying to jumpto some address
the dispatcher processes crashed.

When a new process connects to the dispatcher port
but it does not intend to be a full functional client,
the dispatcher server will assign a temporary `client`
structure to it with name=None, however, the checks
in server loop crashed as it tries to lowercase the
None object.